cam/requirements.txt
requests
numpy
python-vlc
imageio
imageio-ffmpeg
pillow
python-dotenv
flask
rpi-lgpio
psutil
PyYAML
flask_cors